# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Maturity Model API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Skill {
  id: number;
  name: string;
  description: string;
  category: string;
  levelDescriptions: string[];
  levelRequirements: string[];
  levelArtifacts: string[];
  levelRecommendations: string[];
}

export type TeamCriticality =
  (typeof TeamCriticality)[keyof typeof TeamCriticality];

export const TeamCriticality = {
  MC: "MC",
  "BC+": "BC+",
  BC: "BC",
  BO: "BO",
  OP: "OP",
} as const;

export type TeamAssessmentStatus =
  (typeof TeamAssessmentStatus)[keyof typeof TeamAssessmentStatus];

export const TeamAssessmentStatus = {
  planned: "planned",
  in_progress: "in_progress",
  completed: "completed",
  on_hold: "on_hold",
} as const;

export interface Team {
  id: number;
  name: string;
  description: string;
  criticality?: TeamCriticality;
  overallLevel: number;
  assessmentStatus: TeamAssessmentStatus;
  lastAssessedAt?: string | null;
  createdAt: string;
  deletedAt?: string | null;
}

export interface TeamSkillLevel {
  teamId: number;
  skillId: number;
  level: number;
}

export type TeamDetailAssessmentStatus =
  (typeof TeamDetailAssessmentStatus)[keyof typeof TeamDetailAssessmentStatus];

export const TeamDetailAssessmentStatus = {
  planned: "planned",
  in_progress: "in_progress",
  completed: "completed",
  on_hold: "on_hold",
} as const;

export interface TeamSkillWithInfo {
  skillId: number;
  skillName: string;
  skillCategory: string;
  skillDescription: string;
  levelDescriptions: string[];
  levelRequirements: string[];
  levelArtifacts: string[];
  levelRecommendations: string[];
  level: number;
}

export interface TeamDetail {
  id: number;
  name: string;
  description: string;
  overallLevel: number;
  assessmentStatus: TeamDetailAssessmentStatus;
  lastAssessedAt?: string | null;
  createdAt: string;
  skillLevels: TeamSkillWithInfo[];
}

export interface Artifact {
  id: number;
  teamId: number;
  skillId: number;
  name: string;
  link?: string | null;
  note?: string | null;
  createdAt: string;
}

export type CreateTeamRequestCriticality =
  (typeof CreateTeamRequestCriticality)[keyof typeof CreateTeamRequestCriticality];

export const CreateTeamRequestCriticality = {
  MC: "MC",
  "BC+": "BC+",
  BC: "BC",
  BO: "BO",
  OP: "OP",
} as const;

export type CreateTeamRequestTeamType =
  (typeof CreateTeamRequestTeamType)[keyof typeof CreateTeamRequestTeamType];

export const CreateTeamRequestTeamType = {
  product: "product",
  platform: "platform",
  service: "service",
} as const;

export interface CreateTeamRequest {
  name: string;
  description: string;
  orgUnitId?: number | null;
  criticality?: CreateTeamRequestCriticality;
  teamType?: CreateTeamRequestTeamType;
}

export type UpdateTeamRequestCriticality =
  (typeof UpdateTeamRequestCriticality)[keyof typeof UpdateTeamRequestCriticality];

export const UpdateTeamRequestCriticality = {
  MC: "MC",
  "BC+": "BC+",
  BC: "BC",
  BO: "BO",
  OP: "OP",
} as const;

export type UpdateTeamRequestTeamType =
  (typeof UpdateTeamRequestTeamType)[keyof typeof UpdateTeamRequestTeamType];

export const UpdateTeamRequestTeamType = {
  product: "product",
  platform: "platform",
  service: "service",
} as const;

export interface UpdateTeamRequest {
  name: string;
  description: string;
  orgUnitId?: number | null;
  criticality?: UpdateTeamRequestCriticality;
  teamType?: UpdateTeamRequestTeamType;
}

export type UpdateTeamStatusRequestAssessmentStatus =
  (typeof UpdateTeamStatusRequestAssessmentStatus)[keyof typeof UpdateTeamStatusRequestAssessmentStatus];

export const UpdateTeamStatusRequestAssessmentStatus = {
  planned: "planned",
  in_progress: "in_progress",
  completed: "completed",
  on_hold: "on_hold",
} as const;

export interface UpdateTeamStatusRequest {
  assessmentStatus: UpdateTeamStatusRequestAssessmentStatus;
}

export interface UpdateSkillLevelRequest {
  /**
   * @minimum 0
   * @maximum 3
   */
  level: number;
}

export interface CreateArtifactRequest {
  name: string;
  link?: string | null;
  note?: string | null;
}

export interface OrgUnitNode {
  id: number;
  name: string;
  description?: string | null;
  parentId?: number | null;
  children: OrgUnitNode[];
  teamCount: number;
}

export interface CreateOrgUnitRequest {
  name: string;
  description?: string | null;
  parentId?: number | null;
}

export interface UpdateOrgUnitRequest {
  name?: string;
  description?: string | null;
  parentId?: number | null;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}