# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Maturity Model API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: teams
    description: Team operations
  - name: skills
    description: Skill operations
  - name: artifacts
    description: Team skill artifact operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /teams:
    get:
      operationId: getTeams
      tags: [teams]
      summary: Get all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/deleted:
    get:
      operationId: getDeletedTeams
      tags: [teams]
      summary: Get soft-deleted teams
      responses:
        "200":
          description: List of deleted teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team with skill levels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetail"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTeam
      tags: [teams]
      summary: Update team name and description
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Soft-delete team (keeps record, sets deletedAt)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teams/{teamId}/restore:
    post:
      operationId: restoreTeam
      tags: [teams]
      summary: Restore a soft-deleted team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}/status:
    patch:
      operationId: updateTeamStatus
      tags: [teams]
      summary: Update assessment status of a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamStatusRequest"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}/skills/{skillId}:
    put:
      operationId: updateSkillLevel
      tags: [teams]
      summary: Update skill level for a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSkillLevelRequest"
      responses:
        "200":
          description: Updated skill level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSkillLevel"
  /teams/{teamId}/skills/{skillId}/artifacts:
    get:
      operationId: getArtifacts
      tags: [artifacts]
      summary: Get artifacts for a team skill
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
    post:
      operationId: createArtifact
      tags: [artifacts]
      summary: Add artifact to a team skill
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtifactRequest"
      responses:
        "201":
          description: Created artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
  /teams/{teamId}/skills/{skillId}/artifacts/{artifactId}:
    delete:
      operationId: deleteArtifact
      tags: [artifacts]
      summary: Delete an artifact
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
        - name: artifactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /org-units:
    get:
      operationId: getOrgUnits
      tags: [org-units]
      summary: Get organizational units tree
      responses:
        "200":
          description: Organizational units tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgUnitNode"
    post:
      operationId: createOrgUnit
      tags: [org-units]
      summary: Create organizational unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgUnitRequest"
      responses:
        "201":
          description: Created organizational unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnitNode"
  /org-units/{id}:
    put:
      operationId: updateOrgUnit
      tags: [org-units]
      summary: Update organizational unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgUnitRequest"
      responses:
        "200":
          description: Updated organizational unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnitNode"
    delete:
      operationId: deleteOrgUnit
      tags: [org-units]
      summary: Delete organizational unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /skills:
    get:
      operationId: getSkills
      tags: [skills]
      summary: Get all QA skills definition
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        levelDescriptions:
          type: array
          items:
            type: string
        levelRequirements:
          type: array
          items:
            type: string
        levelArtifacts:
          type: array
          items:
            type: string
        levelRecommendations:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - category
        - levelDescriptions
        - levelRequirements
        - levelArtifacts
        - levelRecommendations
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        criticality:
          type: string
          enum: [MC, BC+, BC, BO, OP]
          default: BC
        overallLevel:
          type: integer
        assessmentStatus:
          type: string
          enum: [planned, in_progress, completed, on_hold]
        lastAssessedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - description
        - overallLevel
        - assessmentStatus
        - createdAt
    TeamSkillLevel:
      type: object
      properties:
        teamId:
          type: integer
        skillId:
          type: integer
        level:
          type: integer
      required:
        - teamId
        - skillId
        - level
    TeamDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        overallLevel:
          type: integer
        assessmentStatus:
          type: string
          enum: [planned, in_progress, completed, on_hold]
        lastAssessedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        skillLevels:
          type: array
          items:
            $ref: "#/components/schemas/TeamSkillWithInfo"
      required:
        - id
        - name
        - description
        - overallLevel
        - assessmentStatus
        - createdAt
        - skillLevels
    TeamSkillWithInfo:
      type: object
      properties:
        skillId:
          type: integer
        skillName:
          type: string
        skillCategory:
          type: string
        skillDescription:
          type: string
        levelDescriptions:
          type: array
          items:
            type: string
        levelRequirements:
          type: array
          items:
            type: string
        levelArtifacts:
          type: array
          items:
            type: string
        levelRecommendations:
          type: array
          items:
            type: string
        level:
          type: integer
      required:
        - skillId
        - skillName
        - skillCategory
        - skillDescription
        - levelDescriptions
        - levelRequirements
        - levelArtifacts
        - levelRecommendations
        - level
    Artifact:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        skillId:
          type: integer
        name:
          type: string
        link:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - teamId
        - skillId
        - name
        - createdAt
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        orgUnitId:
          type: integer
          nullable: true
        criticality:
          type: string
          enum: [MC, BC+, BC, BO, OP]
          default: BC
        teamType:
          type: string
          enum: [product, platform, service]
          default: service
      required:
        - name
        - description
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        orgUnitId:
          type: integer
          nullable: true
        criticality:
          type: string
          enum: [MC, BC+, BC, BO, OP]
        teamType:
          type: string
          enum: [product, platform, service]
      required:
        - name
        - description
    UpdateTeamStatusRequest:
      type: object
      properties:
        assessmentStatus:
          type: string
          enum: [planned, in_progress, completed, on_hold]
      required:
        - assessmentStatus
    UpdateSkillLevelRequest:
      type: object
      properties:
        level:
          type: integer
          minimum: 0
          maximum: 3
      required:
        - level
    CreateArtifactRequest:
      type: object
      properties:
        name:
          type: string
        link:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required:
        - name
    OrgUnitNode:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/OrgUnitNode"
        teamCount:
          type: integer
      required:
        - id
        - name
        - children
        - teamCount
    CreateOrgUnitRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
      required:
        - name
    UpdateOrgUnitRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success