# lib/api-zod/src/generated/types/createTeamRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Maturity Model API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTeamRequestCriticality } from "./createTeamRequestCriticality";
import type { CreateTeamRequestTeamType } from "./createTeamRequestTeamType";

export interface CreateTeamRequest {
  name: string;
  description: string;
  orgUnitId?: number | null;
  criticality?: CreateTeamRequestCriticality;
  teamType?: CreateTeamRequestTeamType;
}