# lib/api-zod/src/generated/types/orgUnitNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Maturity Model API
 * OpenAPI spec version: 0.1.0
 */

export interface OrgUnitNode {
  id: number;
  name: string;
  description?: string | null;
  parentId?: number | null;
  children: OrgUnitNode[];
  teamCount: number;
}