# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Maturity Model API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamAssessmentStatus } from "./teamAssessmentStatus";
import type { TeamCriticality } from "./teamCriticality";

export interface Team {
  id: number;
  name: string;
  description: string;
  criticality?: TeamCriticality;
  overallLevel: number;
  assessmentStatus: TeamAssessmentStatus;
  lastAssessedAt?: Date | null;
  createdAt: Date;
  deletedAt?: Date | null;
}