# lib/api-zod/src/generated/types/updateTeamRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Maturity Model API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTeamRequestCriticality } from "./updateTeamRequestCriticality";
import type { UpdateTeamRequestTeamType } from "./updateTeamRequestTeamType";

export interface UpdateTeamRequest {
  name: string;
  description: string;
  orgUnitId?: number | null;
  criticality?: UpdateTeamRequestCriticality;
  teamType?: UpdateTeamRequestTeamType;
}